# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PROSPECTIVE Web API",
    "description": "REST + WebSocket API for cerebral aneurysm segmentation, detection, morphometry, perforator risk assessment and stent planning.\n\nAll linear dimensions are in **millimetres (mm)**. All HU values follow the standard Hounsfield scale.",
    "contact": {
      "name": "SkullApp — Laboratorio de Imagen Médica",
      "email": "ingprospective@skullapp.tech"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/detect/{session_id}": {
      "delete": {
        "tags": [
          "detection"
        ],
        "summary": "Clear detected candidates and their morphometry",
        "description": "Removes the candidate dome meshes, the detection state and the morphometry derived from them — including a manually marked neck plane, which is otherwise reapplied to every later measurement. Use it to start the analysis over on an edited mesh. Idempotent.",
        "operationId": "clear_detection_api_detect__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clear Detection Api Detect  Session Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "detection"
        ],
        "summary": "Detect aneurysm candidates in the segmented mesh",
        "description": "Runs the v6 aneurysm detector on the stored vessel-tree mesh.\n\n**Algorithm:** Gaussian curvature thresholding → connected-component analysis → 3 hard shape gates (positive_gauss_frac, compactness, sphericity) → composite score.\n\nEach candidate is written as an isolated `.vtp` mesh and its URL is returned so vtk.js can render individual domes.\n\n**Prerequisite:** `POST /segment` must be called first.",
        "operationId": "detect_aneurysm_api_detect__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AneurysmDetectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/clips/plan": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Compute clip placement plan",
        "description": "Builds a real 3D mesh of each placed clip, checks collision against the segmented vessel with vtkCollisionDetectionFilter, and estimates neck coverage from the intersection of the clips with the neck plane. Returns the combined clip mesh URL (.vtp) for the 3D viewer.",
        "operationId": "plan_clips_api_clips_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipPlanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/coils/plan": {
      "post": {
        "tags": [
          "coils"
        ],
        "summary": "Compute coil embolization plan",
        "description": "Accepts a list of coil placements and estimates the resulting packing density based on coil wire volumes and the aneurysm sac volume (from session morphometry). Returns packing density and estimated angiographic occlusion grade.\n\nTarget: ≥ 20% packing density for durable occlusion (Sluzewski AJNR 2004).",
        "operationId": "plan_coils_api_coils_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoilPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoilPlanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/centerline/{session_id}": {
      "post": {
        "tags": [
          "centerline"
        ],
        "summary": "Extract the vessel centreline between two points",
        "description": "Computes the medial-axis centreline of the segmented vessel between a source and a target point (voxelisation + EDT + Dijkstra on the distance-weighted grid). Returns a tube mesh of varying radius plus clinical metrics: arc length, tortuosity and vessel diameter statistics.",
        "operationId": "compute_centerline_api_centerline__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CenterlineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterlineResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "centerline"
        ],
        "summary": "Discard the extracted centreline",
        "description": "Deletes the centreline tube, its cached medial-axis points and any stent deployed along it. The centreline-guided stent is built from those points, so leaving it behind would keep a device following a centreline that no longer exists — which is why both go in one call.\n\nIdempotent: succeeds with `had_centerline=false` when nothing was extracted.",
        "operationId": "clear_centerline_api_centerline__session_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterlineClearResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mesh-crop/{session_id}": {
      "post": {
        "tags": [
          "mesh-edit"
        ],
        "summary": "Crop the vessel mesh to a box/sphere ROI",
        "description": "Non-destructive ROI crop of the working vessel mesh. `mode='sphere'` keeps geometry within `radius` of the picked centre; `mode='box'` uses an axis-aligned box (`half_size` per axis, or `radius` as a cube half-side). `invert=true` removes the ROI instead (useful to delete a bone/noise blob). The result overwrites `vessel_tree.vtp`; re-run segmentation to restore.",
        "operationId": "mesh_crop_api_mesh_crop__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeshCropRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeshCropResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mesh-restore/{session_id}": {
      "get": {
        "tags": [
          "mesh-edit"
        ],
        "summary": "How far the vessel mesh can be rolled back",
        "description": "Reports how many crop/grow edits are still undoable for this session, so the mesh-tools panel can enable «Deshacer» and «Restaurar malla original» instead of guessing (a resumed session starts with no client-side history).",
        "operationId": "mesh_history_api_mesh_restore__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeshHistoryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mesh-edit"
        ],
        "summary": "Undo a mesh edit / restore the segmented mesh",
        "description": "`scope='undo'` puts back the mesh as it was before the last ROI crop or grow-from-seeds. `scope='original'` restores the mesh the segmentation produced, discarding every interactive edit.\n\nBoth are cheap file operations — no re-segmentation. Everything derived from the mesh (candidates, morphometry, centreline) must be re-run, which is what the frontend does after calling this.",
        "operationId": "mesh_restore_api_mesh_restore__session_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeshRestoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeshRestoreResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AneurysmDetectionResult": {
        "properties": {
          "found": {
            "type": "boolean",
            "title": "Found",
            "description": "True when at least one candidate was found"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/AneurysmCandidate"
            },
            "type": "array",
            "title": "Candidates",
            "description": "All detected aneurysm candidates, ordered by confidence (descending)"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/DetectionDiagnostics",
            "description": "Region counts per rejection reason — explains an empty result"
          }
        },
        "type": "object",
        "required": [
          "found"
        ],
        "title": "AneurysmDetectionResult",
        "description": "Result of the aneurysm detection step."
      },
...
      "ClipPlanRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "placements": {
            "items": {
              "$ref": "#/components/schemas/ClipPlacement"
            },
            "type": "array",
            "title": "Placements",
            "description": "All clips to place (replaces the current plan)"
          },
          "trajectory_entry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position3D"
              },
              {
                "type": "null"
              }
            ],
            "description": "Surgical approach entry point in patient space (mm)"
          },
          "trajectory_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Position3D"
              },
              {
                "type": "null"
              }
            ],
            "description": "Surgical approach target point in patient space (mm)"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "placements"
        ],
        "title": "ClipPlanRequest",
        "description": "Request to add or update a clip in the planning."
      },
...
      "CoilPlanRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "placements": {
            "items": {
              "$ref": "#/components/schemas/CoilPlacement"
            },
            "type": "array",
            "title": "Placements",
            "description": "All coils placed inside the aneurysm sac"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "placements"
        ],
        "title": "CoilPlanRequest",
        "description": "Request to update the coil embolization plan."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}